Feature: Updating pull energy data file and creat edownloadlink

+++ doesGermanyBurnMoreCoal/pull_energy_charts_data.py
@@ -12,19 +12,11 @@ def pull_data_from_url(url):
         json_data = {}
     return json_data
 
-
-def create_filename(folder, query_string):
-    filename = os.path.join(folder, query_string)
-    return filename
-
-
-def create_url(search_string, base_url):
-    url = os.path.join(base_url, search_string)
-    return url
-
-
-def write_json_data_to_file(json_data, filename):
-    with open (filename, "w") as f:
+def write_json_data_to_file(folder, filename, json_data):
+    if not os.path.exists(folder):
+        raise FileNotFoundError(f"{folder}/ - Folder does not exist.")
+    export_path = os.path.join(folder, filename)
+    with open (export_path, "w") as f:
         json.dump(json_data, f)
 
 
@@ -32,31 +24,32 @@ def create_query_string(intervall_query_string, date_query_string):
     search_string = intervall_query_string + "_" + str(date_query_string) + ".json"
     return search_string
 
+def read_search_urls_list_from_file(search_string_file):
+    with open(search_string_file, "r") as f:
+        search_urls = json.load(f)
 
-def download_energy_data(url):
-    return
+    return search_urls
+
+
+def create_export_filename(url):
+    split_url = url.split("/")
+    export_filename = split_url[5] + "_" + split_url[-1]
+    return export_filename
 
 
 def main():
-    year_list = list(range(1990, 2023))
-    domain = "https://www.energy-charts.info"
-    base_url = "https://www.energy-charts.info/charts/power/data/de/"
-    # /charts/power/raw_data/de/year_2010.json
-    intervall_query_string = "year"
-    search_string_list = []
-    folder = "energy_data"
-
-    for year in year_list:
-        search_string_list.append(create_query_string(intervall_query_string, year))
-
-    for search_string in search_string_list:
-        filename = create_filename("", search_string)
-        url = create_url(search_string, base_url)
-        print(url)
-        energy_data = pull_data_from_url(url)
-
-        if len(energy_data) > 0:
-            write_json_data_to_file(energy_data, filename)
+    url_file = "search_urls.json"
+    folder = "data"
+    search_urls = read_search_urls_list_from_file(url_file)
+    for url in search_urls:
+        filename = create_export_filename(url)
+        try:
+            energy_data = pull_data_from_url(url)
+            if len(energy_data) > 1:
+                write_json_data_to_file(folder, filename, energy_data)
+        except Exception as e:
+            print("skipping url", url)
+            print(e.message)
 
 if __name__ == "__main__":
     main()
